+++ src/gdetect/api.py
@@ -65,20 +65,15 @@ from .exceptions import (
     BadLayoutError,
 )
 from .stream import StreamReader
+from .consts import GDETECT_USER_AGENT, EXPORT_LAYOUTS, EXPORT_FORMATS
 
 logger = get_logger()
 
 BASE_ENDPOINT = "/api/lite/v2"
 
-UUID_PATTERN = re.compile(
-    r"[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}"
-)
+UUID_PATTERN = re.compile(r"[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}")
 SHA256_PATTERN = re.compile(r"[0-9a-f]{64}")
-TOKEN_PATTERN = re.compile(
-    r"[0-9a-f]{8}-[0-9a-f]{8}-[0-9a-f]{8}-[0-9a-f]{8}-[0-9a-f]{8}"
-)
-GDETECT_USER_AGENT = "py-gdetect/0.7.0"
-
+TOKEN_PATTERN = re.compile(r"[0-9a-f]{8}-[0-9a-f]{8}-[0-9a-f]{8}-[0-9a-f]{8}-[0-9a-f]{8}")
 
 @dataclass
 class Status:
@@ -408,9 +403,7 @@ class Client:
         token = resp.get("token", "")
         if token == "":
             raise MissingTokenError()
-        return urllib.parse.urljoin(
-            self.base_url, f"/expert/en/analysis-redirect/{token}"
-        )
+        return urllib.parse.urljoin(self.base_url, f"/expert/en/analysis-redirect/{token}")
 
     def extract_expert_url(self, resp: dict) -> str:
         """Extract expert view from response.
@@ -425,13 +418,9 @@ class Client:
         sid = resp.get("sid", "")
         if sid == "":
             raise MissingSIDError()
-        return urllib.parse.urljoin(
-            self.base_url, f"/expert/en/analysis/advanced/{sid}"
-        )
+        return urllib.parse.urljoin(self.base_url, f"/expert/en/analysis/advanced/{sid}")
 
-    def export_result(
-        self, uuid: str, format: str, layout: str, full: bool = False
-    ) -> bytes:
+    def export_result(self, uuid: str, format: str, layout: str, full: bool = False) -> bytes:
         """Export analysis result with the requested layout and format
 
         Args:
@@ -468,11 +457,11 @@ class Client:
         return resp.content
 
     def _check_export_format(self, format: str):
-        if format not in ["misp", "stix", "json", "pdf", "markdown", "csv"]:
+        if format not in EXPORT_FORMATS:
             raise BadExportFormatError
 
     def _check_layout(self, layout: str):
-        if layout not in ["fr", "en"]:
+        if layout not in EXPORT_LAYOUTS:
             raise BadLayoutError
 
     def _check_uuid(self, uuid: str):
@@ -491,9 +480,7 @@ class Client:
         if not TOKEN_PATTERN.match(self.token):
             raise BadAuthenticationTokenError("bad token format")
 
-    def _request(
-        self, method: str, url: str, headers: dict = {}, timeout: float = 30, **kwargs
-    ) -> requests.Response:
+    def _request(self, method: str, url: str, headers: dict = {}, timeout: float = 30, **kwargs) -> requests.Response:
         """Process a request to URL with given method and params.
 
         This function execute the request to the URL with `requests` library.
@@ -511,9 +498,7 @@ class Client:
         # set auth token if not provided
         headers["X-Auth-Token"] = headers.get("X-Auth-Token", self.token)
         headers["User-Agent"] = GDETECT_USER_AGENT
-        resp = requests.request(
-            method, url, verify=self.verify, headers=headers, timeout=timeout, **kwargs
-        )
+        resp = requests.request(method, url, verify=self.verify, headers=headers, timeout=timeout, **kwargs)
         code = resp.status_code
         if code != 200:
             raise compute_exception_from_response(resp)