+++ src/gdetect/consts.py
@@ -1,0 +1,7 @@
+"""
+gdetect constants
+"""
+
+GDETECT_USER_AGENT = "py-gdetect/0.8.0"
+EXPORT_FORMATS = ["misp", "stix", "json", "pdf", "markdown", "csv"]
+EXPORT_LAYOUTS = ["en", "fr"]