+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-* Export command : allow to export submisssion result in a given format and layout, full or summarized
+* Export command: allow to export submission result in a given format and layout, full or summarized
 
 ## v0.7.0